+++ src/util.rs
@@ -1,0 +1,2 @@
+mod generate_password;
+pub use generate_password::generate_password;