+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this
 project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 2025-07-01
+
+Add flexible possibility to generate passwords.
+
+Use dedicated action to close modals, allowing for keeping the right state of the main UI.
+
+Show warnings in log in orange rather than yellow.
+
 ## [1.1.1] - 2025-05-23
 
 Deactivate the tabs and the create button if a bundle or a document are being modified.